Reshuffling package structure
