+++ trigger-support/pom.xml
@@ -16,6 +16,16 @@
             <groupId>gg.xp</groupId>
             <artifactId>xivsupport</artifactId>
         </dependency>
+        <dependency>
+            <groupId>gg.xp</groupId>
+            <artifactId>testutils-xiv</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.hamcrest</groupId>
+            <artifactId>hamcrest-all</artifactId>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
 </project>
