+++ trigger-support/src/test/java/gg/xp/xivsupport/events/triggers/seq/SequentialTriggerTest.java
