+++ triggers/triggers-ew/pom.xml
@@ -3,22 +3,13 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>triggevent</artifactId>
+        <artifactId>triggers</artifactId>
         <groupId>gg.xp</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>plugins</artifactId>
+    <artifactId>triggers-ew</artifactId>
 
-    <packaging>pom</packaging>
-    <modules>
-        <module>dot-tracker</module>
-        <module>cd-tracker</module>
-        <module>job-overlays</module>
-        <module>personal-cast-bar-overlay</module>
-        <module>skill-hit-overlay</module>
-        <module>custom-party-overlay</module>
-    </modules>
 
 </project>
