+++ triggers/triggers-ew/src/main/java/gg/xp/xivsupport/events/triggers/duties/Pandamonium/P8S2.java
