+++ triggers/triggers-ew/src/test/java/gg/xp/xivsupport/events/triggers/endwalker/extremes/EX3Test.java
