+++ triggers/triggers-sb/src/test/java/gg/xp/xivsupport/events/triggers/uwu/UWUTest.java
